% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjlist_to_matrix.R
\name{adjlist_to_matrix}
\alias{adjlist_to_matrix}
\title{Convert adjacency list into matrix}
\usage{
adjlist_to_matrix(input_df)
}
\arguments{
\item{input_df}{A tibble with NxN observations, such that every row is a
pairwise combination of two nodes (`from` and `to`), and a column called
`edge` encodes whether there is a relationship between them.}
}
\value{
A square NxN matrix, where N=number of nodes.
}
\description{
An adjacency contains NxN observations that can be alternatively represented
as an NxN square matrix. This function performs that conversion.
}
\examples{
`\%>\%` <- magrittr::`\%>\%`
successr::karate \%>\%
    tidygraph::as_tbl_graph(directed = F) \%>\%
    graph_to_adjlist() \%>\%
    adjlist_to_matrix()

}
