% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_nothing.R
\name{do_nothing}
\alias{do_nothing}
\title{Does nothing}
\usage{
do_nothing(input)
}
\arguments{
\item{input}{Obligatory input.}
}
\value{
Returns the input.
}
\description{
Does nothing. Exists for developer convenience, but pointless for users.
Allows for NAMESPACE imports to be collected in one place.
Should not be exported.
}
