% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_experiment.R
\name{generate_experiment}
\alias{generate_experiment}
\title{Simulate a basic experiment}
\usage{
generate_experiment(input_tidygraph, n_reps)
}
\arguments{
\item{input_tidygraph}{A `tbl_graph` representation from `tidygraph`.
This function may not work with other kinds of network representations,
e.g. from `igraph`. Make sure that your input graph is either directed
or undirected, as you'd like it.}

\item{n_reps}{Number of repetitions.}
}
\description{
Generate N pairwise observations of each relationship, like in a (balanced)
laboratory experiment.
}
\examples{
`\%>\%` <- magrittr::`\%>\%`
successr::karate \%>\%
    tidygraph::as_tbl_graph(directed = F) \%>\%
    generate_experiment(2)

}
