% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_random_walk.R
\name{generate_random_walk}
\alias{generate_random_walk}
\title{Random walk around a graph}
\usage{
generate_random_walk(input_tidygraph, n_obs, start_here = NULL)
}
\arguments{
\item{input_tidygraph}{A `tbl_graph` representation from `tidygraph`.
This function may not work with other kinds of network representations,
e.g. from `igraph`. Make sure that your input graph is either directed
or undirected, as you'd like it.}

\item{n_obs}{Number of observations (i.e., steps in the random walk).}

\item{start_here}{Numeric node ID. By default, picks a random starting point.}
}
\value{
A tibble containing all observations.
}
\description{
Generate a random walk around a graph. Does *not* factor edge weights into
the walk (i.e., will transition to any successor node with equal
probability).
}
\examples{
`\%>\%` <- magrittr::`\%>\%`
successr::karate \%>\%
    tidygraph::as_tbl_graph(directed = F) \%>\%
    generate_random_walk(100)

}
