% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuccessPlan.R
\name{load_daydata}
\alias{load_daydata}
\title{load_daydata}
\usage{
load_daydata(data_path = file.path(Sys.getenv("R_SUCCESS"), "DayData.txt"))
}
\arguments{
\item{data_path}{Path to your DayData.txt file. The default path is 
\code{Sys.getenv("R_SUCCESS")}.}
}
\value{
data.frame
}
\description{
Load the DayData.txt dataset.
}
