% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_to_adjlist.R
\name{matrix_to_adjlist}
\alias{matrix_to_adjlist}
\title{Convert matrix into adjacency list}
\usage{
matrix_to_adjlist(input_matrix, edge_col_name = "edge")
}
\arguments{
\item{input_matrix}{A square NxN matrix, where N=number of nodes.}

\item{edge_col_name}{The name of the column encoding the relation between
two given nodes. This is probably `edge` for adjacency matrices, or
`successor_value` for matrices encoding successor relations.}
}
\value{
A tibble with NxN observations, such that every row is a pairwise
    combination of two nodes (`from` and `to`), and a column encoding the
    relationship between them.
}
\description{
An NxN square matrix can be alternatively represented as an adjacency list
containing NxN observations. This function performs that conversion.
}
\examples{
`\%>\%` <- magrittr::`\%>\%`
successr::karate \%>\%
    tidygraph::as_tbl_graph(directed = F) \%>\%
    graph_to_adjlist() \%>\%
    adjlist_to_matrix() \%>\%
    matrix_to_adjlist()

}
