% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuccessPlan.R
\name{success_summary}
\alias{success_summary}
\title{success_summary}
\usage{
success_summary(year = NULL, more_than = c(3, 5))
}
\arguments{
\item{year}{Year to report; default = NULL to report for all available years.}

\item{more_than}{A vector of integers to report number of days worker for more than x
hours.}
}
\value{

}
\description{
Provides summary statistics for time worked by year.
}
