% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{user_col_to_string}
\alias{user_col_to_string}
\title{Utilities}
\usage{
user_col_to_string(user_col_in_double_brackets)
}
\description{
Utility functions for developer convenience, not intended for users.
Also allows for NAMESPACE imports to be collected in one place.
Should not be exported.
}
