% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_successor.R
\name{initialize_successor}
\alias{initialize_successor}
\title{Initialize successor values}
\usage{
initialize_successor(input_tidygraph, initialize_with = 0)
}
\arguments{
\item{input_tidygraph}{A `tbl_graph` representation from `tidygraph`.
This function will not work with other kinds of network representations,
e.g. from `igraph`. Make sure that your input graph is either directed
or undirected, as you'd like it.}

\item{initialize_with}{A scalar. Defaults to `0`.}
}
\value{
A square NxN matrix, where N=number of nodes.
}
\description{
The successor counts must be initialized to some value. All things equal,
the value `0` tends to be pretty good.
}
\examples{
`\%>\%` <- magrittr::`\%>\%`
successr::karate \%>\%
    tidygraph::as_tbl_graph(directed = F) \%>\%
    initialize_successor()

}
