% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learn_from_param_sweep.R
\name{learn_from_param_sweep}
\alias{learn_from_param_sweep}
\title{Learn successor values from observations (many parameters)}
\usage{
learn_from_param_sweep(
  successor_matrix,
  observations,
  alphas,
  gammas,
  bidirectional = FALSE
)
}
\arguments{
\item{successor_matrix}{A square matrix created by `initialize_successor`.}

\item{observations}{A tibble with observations `from` a node `to` another.}

\item{alphas}{Vector corresponding to the learning rate bound in [0, 1].}

\item{gammas}{Vector corresponding to the lookaround horizon bound in [0, 1).}

\item{bidirectional}{Logical. Defaults to `FALSE`, which means that only
the `from-to` relationship gets updated. If set to `TRUE`, this function
will also update the `to-from` relationship.}
}
\value{
A dataframe with updated successor values, given the observations.
}
\description{
After observing relations between nodes (i.e., transitioning between
states), update the successor values. This function learns from many
observations, and updates according to many parameter values. To update the
successor values from a single observation, use `update_successor` instead.
To update successor values from a single set of parameters, use
`learn_from_observations` instead.
}
\examples{
`\%>\%` <- magrittr::`\%>\%`
karate_graph <- successr::karate \%>\%
    tidygraph::as_tbl_graph(directed = F)
karate_walk <- karate_graph \%>\%
    generate_random_walk(1000)
karate_graph \%>\%
    initialize_successor() \%>\%
    learn_from_param_sweep(karate_walk, c(0.1, 0.2), c(0, 0.4), TRUE)

}
