% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuccessPlan.R
\name{success_plot}
\alias{success_plot}
\title{Plot work hours for the year}
\usage{
success_plot(data_path = NULL, Year = NULL)
}
\arguments{
\item{data_path}{Path to your DayData.csv file. If you have set the R_SUCCESS
environment variable then leave the argument as NULL.}
}
\description{
Plots the amount of time you have worked each month for the year.
}
