% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuccessPlan.R
\name{success_summary}
\alias{success_summary}
\title{success_summary}
\usage{
success_summary(Year = 2018)
}
\arguments{
\item{Year}{an integer of the year}
}
\value{

}
\description{
Summary statistics of time worked
}
