% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuccessPlan.R
\name{successr}
\alias{successr}
\title{successr}
\usage{
successr(verbose = FALSE, sanitize = FALSE)
}
\arguments{
\item{verbose}{prints out configuration settings.}

\item{sanitize}{clears out unused labels from the Task dropdown menu in the \code{Edit}
window. This issue often happens when you waste time tinkering with button labels.}
}
\description{
A diagnostic tool for tracking work and wasted time.
}
\details{
\code{successr} is a
tool for tracking the amount of time you work (and waste) during the day. At a metaphysical level, time
can be divided into two categories: work time and wasted time. Work time is effectively
used when you focus on at most two relevant projects during the day.
\code{Wasted Time} includes anything that does not fall in \code{Project 1} and/or
\code{Project 2}.  Culprits for \code{Wasted Time} are typically
toilet/cigarette/watercooler excursions, lunch breaks, unannounced visits by the boss,
any meeting, internet surfing, chit chat with colleagues, or courtesy calls by
telemarketers, friends/family, and partner(s).

Clicking \code{Project 1}, \code{Project 2}, and \code{Wasted Time} starts the timer,
and \code{Stop} stops the timer. \code{Report} gives a daily, weekly, monthly breakdown
of work time. \code{Edit} allows you to edit only the time recorded during the day and
the Task (from a drop-down menu).  \code{DayData.csv} allows you to edit the total time
for a given day (see below). 

Upon installation, \code{successr} will setup a data folder (where the data is stored) with a
configuration file, called \code{config.yml} (see \code{\link{config}}). The folder can
be found in \code{Sys.getenv("HOME")}. If you want to specify an alternative folder
path, then you must put in your .Rprofile (see \code{\link{Startup}}) the following
line: \code{Sys.setenv(R_SUCCESS="my/folder/path")}. In my Rprofile set-up, I have
\code{Sys.setenv("~/Dropbox/successr")} so I can sync between my work and home
computers. Labels for \code{Project 1}, \code{Project 2}, \code{Wasted Time} and
the \code{Window Title} can be changed in the \code{config.yml} file. But don't
waste time doing this, just use the default settings.  

Requires \code{GTK} libraries to work that you may (or may not) have to manually install
yourself (which will waste some time). _successr_ is in the
development phase, please report bugs to me by typing in the \code{R} console
\code{packageDescription("successr")}.
}
